+++ packages/app-icons-loader/utils.ts
@@ -1,0 +1,8 @@
+export const debounce = (func: any, wait: number) => {
+  let timeout: any;
+  return function (this: any, ...args: any[]) {
+    const context = this;
+    clearTimeout(timeout);
+    timeout = setTimeout(() => func.apply(context, args), wait);
+  };
+};
